+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Break long terms to prevent overflow
 - Improve page heading hierarchy
 - Improve aria labelling
+- Use error message in page title
 
 0.1.1 / 2018-07-10
 ==================